// src/models/DescCen3.ts
export interface DescCen3 {
  IdDescCen3: number;
  nameDescCen3: string | null;
  codeDescCen3: string | null;
}

export interface DescCen3Create {
  nameDescCen3: string | null;
  codeDescCen3: string | null;
}

export interface DescCen3Update {
  nameDescCen3: string | null;
  codeDescCen3: string | null;
}